Multiple bug fixes and changes after code reviews: - Reviewed r134, minor change. - Reviewed r137, minor changes:   * Columns can now only be removed when they're not created by LOVD, like in LOVD 2.0.   * Standardized the columns?delete a bit, removed old LOVD 2.0 code from columns.php. - Reviewed r138, minor improvements added. - Replaced some old lovd_queryDB() calls with PDO calls in class/objects.php. - Fixed bug; The email sent after a submitter registers, did not contain a proper link directly to the user's profile. - Fixed notice directly after a submitter registers. - The email forwarded to the admin after a submitter registers, now has a proper Reply-To header. - Fixed bug; "Not authorized" messages would include the header even if the "clean" header was already included before. - Fixed bug; lovd_isAuthorized() returned false for users lower than managers when authorizing for variants, individuals, phenotypes and screenings. - Slightly optimized lovd_isAuthorized(). - Fixed bug; lovd_matchURL() didn't correctly parse the given URLs that could result in false positives or false negatives. - Fixed problems with the viewList checkboxes; all checkboxes were enabled when removing a filter, after you had selected the full filter subset. - Fixed bug; It was not possible to activate the "Forward messages to database admin" or "Enable submitters to change data" settings. - Fixed bug; On servers where MySQL had the ONLY_FULL_GROUP_BY setting enabled, custom columns could not be edited nor enabled. - Updated variant viewLists; the genomic variants viewList no longer shows the internal type column and the variant effect column was moved to the left. The /variants/GENE viewList now correctly loads the columns active for that gene, in the correct order. - HGVS columns somehow not enabled had a "Delete column" link that led to a "Hack Attempt!" page. - The authorization unit test was expanded with authorization on variants for curators, collaborators and submitters. - Standardized the links to owners in various viewLists and viewEntries.
git-svn-id: https://humgenprojects.lumc.nl/svn/LOVD3/trunk@140 a7477d18-a37e-46de-aef1-57953c184492

+++ src/columns.php
@@ -4,7 +4,7 @@
  * LEIDEN OPEN VARIATION DATABASE (LOVD)
  *
  * Created     : 2010-03-04
- * Modified    : 2012-03-13
+ * Modified    : 2012-03-19
  * For LOVD    : 3.0-beta-03
  *
  * Copyright   : 2004-2012 Leiden University Medical Center; http://www.LUMC.nl/
@@ -109,41 +109,26 @@ if (!empty($_PATH_ELEMENTS[2]) && !ACTION) {
         } else {
             $sNavigation = '<A style="color : #999999;">Enable column</A>';
         }
-        // Remove column.
+        // Disable column.
         if ($zData['active'] && !$zData['hgvs']) {
             $sNavigation .= ' | <A href="columns/' . $zData['id'] . '?remove">Disable column</A>';
-            $sNavigation .= ' | <A style="color : #999999;">Delete column</A>';
         } else {
             $sNavigation .= ' | <A style="color : #999999;">Disable column</A>';
+        }
+        // Delete column.
+        if (!$zData['active'] && !$zData['hgvs'] && (int) $zData['created_by']) {
             $sNavigation .= ' | <A href="columns/' . $zData['id'] . '?delete">Delete column</A>';
+        } else {
+            $sNavigation .= ' | <A style="color : #999999;">Delete column</A>';
         }
         $sNavigation .= ' | <A href="columns/' . $zData['id'] . '?edit">Edit custom data column settings</A>';
         $sNavigation .= ' | <A href="columns/' . $zData['category'] . '?order">Re-order all ' . $zData['category'] . ' columns</A>';
 /*
 
-        // Drop global column.
-        $bSelected = true;
-        if (substr($zData['colid'], 0, 7) == 'Variant') {
-            // Check genes to find if column is active.
-            $aGenes = lovd_getGeneList();
-            foreach ($aGenes as $sSymbol) {
-                list($bSelected) = mysql_fetch_row(mysql_query('SELECT colid FROM ' . TABLEPREFIX . '_' . $sSymbol . '_columns WHERE colid = "' . $zData['colid'] . '"'));
-                if ($bSelected) {
-                    // Column present in this gene.
-                    break;
-                }
-            }
-        } elseif (substr($zData['colid'], 0, 7) == 'Patient') {
-            // Patient column.
-            list($bSelected) = mysql_fetch_row(mysql_query('SELECT colid FROM ' . TABLE_PATIENTS_COLS . ' WHERE colid = "' . $zData['colid'] . '"'));
-        }
-
         if ($zData['created_by'] && !$bSelected) {
             $sNavigation .= ' | <A href="' . $_SERVER['PHP_SELF'] . '?action=edit_colid&amp;edit_colid=' . rawurlencode($zData['colid']) . '">Edit column ID</A>';
-            $sNavigation .= ' | <A href="' . $_SERVER['PHP_SELF'] . '?action=drop&amp;drop=' . rawurlencode($zData['colid']) . '">Delete column</A>';
         } else {
             $sNavigation .= ' | <A style="color : #999999;">Edit column ID</A>';
-            $sNavigation .= ' | <A style="color : #999999;">Delete column</A>';
         }
 */
     }
@@ -691,30 +676,41 @@ if (empty($_PATH_ELEMENTS[1]) && ACTION == 'create') {
     if (empty($_POST['category'])) {
         require ROOT_PATH . 'inc-top.php';
         lovd_printHeader(PAGE_TITLE);
-        $aOptionsList = array('width' => 950);
-        $aOptionsList['options'][0]['onclick']     = '$(\'#optionForm input\').attr(\'value\', \'Individual\'); $(\'#optionForm\').submit();';
-        $aOptionsList['options'][0]['option_text'] = '<B>Information on the individual, not related to disease</B>, not changing over time, such as date of birth';
 
-        $aOptionsList['options'][1]['onclick']     = '$(\'#optionForm input\').attr(\'value\', \'Phenotype\'); $(\'#optionForm\').submit();';
-        $aOptionsList['options'][1]['option_text'] = '<B>Information on the phenotype, related to disease</B>, possibly changing over time, such as blood pressure';
-
-        $aOptionsList['options'][2]['onclick']     = '$(\'#optionForm input\').attr(\'value\', \'Screening\'); $(\'#optionForm\').submit();';
-        $aOptionsList['options'][2]['option_text'] = '<B>Information on the detection of new variants</B>, such as detection technique or laboratory conditions';
-
-        $aOptionsList['options'][3]['onclick']     = '$(\'#optionForm input\').attr(\'value\', \'VariantOnGenome\'); $(\'#optionForm\').submit();';
-        $aOptionsList['options'][3]['option_text'] = '<B>Information on the variant(s) found, in general or on the genomic level</B>, such as restriction site change';
-
-        $aOptionsList['options'][4]['onclick']     = '$(\'#optionForm input\').attr(\'value\', \'VariantOnTranscript\'); $(\'#optionForm\').submit();';
-        $aOptionsList['options'][4]['option_text'] = '<B>Information on the variant(s) found, specific for the transcript level</B>, such as predicted effect on protein level';
+        $aOptionsList =
+             array(
+                    'width' => 950,
+                    'options' =>
+                         array(
+                             array(
+                                    'onclick'     => '$(\'#optionForm input\').attr(\'value\', \'Individual\'); $(\'#optionForm\').submit();',
+                                    'option_text' => '<B>Information on the individual, not related to disease</B>, not changing over time, such as date of birth',
+                                  ),
+                             array(
+                                    'onclick'     => '$(\'#optionForm input\').attr(\'value\', \'Phenotype\'); $(\'#optionForm\').submit();',
+                                    'option_text' => '<B>Information on the phenotype, related to disease</B>, possibly changing over time, such as blood pressure',
+                                  ),
+                             array(
+                                    'onclick'     => '$(\'#optionForm input\').attr(\'value\', \'Screening\'); $(\'#optionForm\').submit();',
+                                    'option_text' => '<B>Information on the detection of new variants</B>, such as detection technique or laboratory conditions',
+                                  ),
+                             array(
+                                    'onclick'     => '$(\'#optionForm input\').attr(\'value\', \'VariantOnGenome\'); $(\'#optionForm\').submit();',
+                                    'option_text' => '<B>Information on the variant(s) found, in general or on the genomic level</B>, such as restriction site change',
+                                  ),
+                             array(
+                                    'onclick'     => '$(\'#optionForm input\').attr(\'value\', \'VariantOnTranscript\'); $(\'#optionForm\').submit();',
+                                    'option_text' => '<B>Information on the variant(s) found, specific for the transcript level</B>, such as predicted effect on protein level',
+                                  ),
+                              ),
+                  );
 
         print('      You\'re about to create a new custom data column. This will allow you to define what kind of information you would like to store in the database. Please note that <I>defining</I> this type of information, does not automatically make LOVD store this information. You will need to <I>enable</I> it after defining it, so it actually gets added to the data entry form.<BR><BR>' . "\n" .
               '      Firstly, please choose what kind of category the new type of data belongs:<BR><BR>' . "\n\n" .
-              '      <FORM id="optionForm" action="' . CURRENT_PATH . '?' . ACTION . '" method="post">' . "\n");
-        print(lovd_buildOptionTable($aOptionsList));
-        print('        <INPUT name="category" type="hidden" value="">' . "\n" .
+              '      <FORM id="optionForm" action="' . CURRENT_PATH . '?' . ACTION . '" method="post">' . "\n" .
+              lovd_buildOptionTable($aOptionsList) .
+              '        <INPUT name="category" type="hidden" value="">' . "\n" .
               '      </FORM>' . "\n\n");
-
-
 
         require ROOT_PATH . 'inc-bot.php';
         exit;
@@ -1233,142 +1229,7 @@ $( function ()
 }
 
 /*
-if ($_GET['action'] == 'drop' && !empty($_GET['drop'])) {
-    // Deleting self-created columns.
-
-// Require manager clearance.
-lovd_requireAUTH(LEVEL_MANAGER);
-
-    $zData = @mysql_fetch_assoc(mysql_query('SELECT * FROM ' . TABLE_COLS . ' WHERE created_by != 0 AND colid = "' . $_GET['drop'] . '"'));
-    if (!$zData) {
-        // Wrong ID, apparently.
-        require ROOT_PATH . 'inc-top.php';
-        lovd_printHeader('setup_columns_manage_defaults', 'LOVD Setup - Manage custom column defaults');
-        lovd_showInfoTable('No such ID!', 'stop');
-        require ROOT_PATH . 'inc-bot.php';
-        exit;
-    }
-
-    $bSelected = true;
-    if (substr($zData['colid'], 0, 7) == 'Variant') {
-        // Check genes to find if column is active.
-        $aGenes = lovd_getGeneList();
-        foreach ($aGenes as $sSymbol) {
-            list($bSelected) = mysql_fetch_row(mysql_query('SELECT colid FROM ' . TABLEPREFIX . '_' . $sSymbol . '_columns WHERE colid = "' . $zData['colid'] . '"'));
-            if ($bSelected) {
-                // Column present in this gene.
-                break;
-            }
-        }
-    } elseif (substr($zData['colid'], 0, 7) == 'Patient') {
-        // Patient column.
-        list($bSelected) = mysql_fetch_row(mysql_query('SELECT colid FROM ' . TABLE_PATIENTS_COLS . ' WHERE colid = "' . $zData['colid'] . '"'));
-    }
-
-    if (!$zData['created_by'] || $bSelected) {
-        require ROOT_PATH . 'inc-top.php';
-        lovd_printHeader('setup_columns_manage_defaults', 'LOVD Setup - Manage custom column defaults');
-        lovd_showInfoTable('Column has been selected, cannot be removed!', 'stop');
-        require ROOT_PATH . 'inc-bot.php';
-        exit;
-    }
-
-    require ROOT_PATH . 'inc-lib-form.php';
-
-    if (isset($_GET['sent'])) {
-        lovd_errorClean();
-
-        // Mandatory fields.
-        $aCheck =
-                 array(
-                        'password' => 'Enter your password for authorization',
-                      );
-
-        foreach ($aCheck as $key => $val) {
-            if (empty($_POST[$key])) {
-                lovd_errorAdd($key, 'Please fill in the \'' . $val . '\' field.');
-            }
-        }
-
-        // User had to enter his/her password for authorization.
-        if ($_POST['password'] && !lovd_verifyPassword($_POST['password'], $_AUTH['password'])) {
-            lovd_errorAdd('password', 'Please enter your correct password for authorization.');
-        }
-
-        if (!lovd_error()) {
-            // Delete the row in the general_columns table.
-            $sQ = 'DELETE FROM ' . TABLE_COLS . ' WHERE colid = "' . $_GET['drop'] . '"';
-            $q = mysql_query($sQ);
-            if (!$q) {
-                $sError = mysql_error(); // Save the mysql_error before it disappears.
-                require ROOT_PATH . 'inc-top.php';
-                lovd_printHeader('setup_columns_manage_defaults', 'LOVD Setup - Manage custom column defaults');
-                lovd_dbFout('ColDrop', $sQ, $sError);
-            }
-
-            // Delete the links in the general_columns2links table.
-            $sQ = 'DELETE FROM ' . TABLE_COLS2LINKS . ' WHERE colid = "' . $_GET['drop'] . '"';
-            $q = mysql_query($sQ);
-            if (!$q) {
-                // Silent error.
-                lovd_writeLog('MySQL:Error', 'ColDrop', 'Custom links could not be removed from ' . $zData['colid']);
-            } else {
-                lovd_writeLog('MySQL:Event', 'ColDrop', 'Custom links successfully removed from ' . $zData['colid']);
-            }
-
-            // Write to log...
-            lovd_writeLog('MySQL:Event', 'ColDrop', $_AUTH['username'] . ' (' . mysql_real_escape_string($_AUTH['name']) . ') successfully deleted column ' . $zData['colid'] . ' (' . mysql_real_escape_string($zData['head_column']) . ')');
-
-            // Thank the user...
-            header('Refresh: 3; url=' . PROTOCOL . $_SERVER['HTTP_HOST'] . $_SERVER['PHP_SELF'] . '?action=view_all');
-
-            require ROOT_PATH . 'inc-top.php';
-            lovd_printHeader('setup_columns_manage_defaults', 'LOVD Setup - Manage custom column defaults');
-            print('      Successfully deleted column "' . $zData['colid'] . '"!<BR><BR>' . "\n\n");
-
-            require ROOT_PATH . 'inc-bot.php';
-            exit;
-
-        } else {
-            // Errors, so the whole lot returns to the form.
-            lovd_magicUnquoteAll();
-
-            // Because we're sending the data back to the form, I need to unset the password fields!
-            unset($_POST['password']);
-        }
-    }
-
-
-
-    require ROOT_PATH . 'inc-top.php';
-    lovd_printHeader('setup_columns_manage_defaults', 'LOVD Setup - Manage custom column defaults');
-
-    lovd_errorPrint();
-
-    // Table.
-    print('      <FORM action="' . $_SERVER['PHP_SELF'] . '?action=' . $_GET['action'] . '&amp;drop=' . rawurlencode($zData['colid']) . '&amp;sent=true" method="post">' . "\n");
-
-    // Array which will make up the form table.
-    $aForm = array(
-                    array('POST', '', '', '50%', '50%'),
-                    array('Permanently deleting column', 'print', $zData['colid'] . ' (' . $zData['head_column'] . ')'),
-                    'skip',
-                    array('Enter your password for authorization', 'password', 'password', 20),
-                    array('', 'submit', 'Delete column permanently'),
-                  );
-    $_MODULES->processForm('SetupColumnsDelete', $aForm);
-    lovd_viewForm($aForm);
-
-    print('</FORM>' . "\n\n");
-
-    require ROOT_PATH . 'inc-bot.php';
-    exit;
-
-
-
-
-
-} elseif ($_GET['action'] == 'edit_colid' && !empty($_GET['edit_colid'])) {
+if ($_GET['action'] == 'edit_colid' && !empty($_GET['edit_colid'])) {
     // Edit specific custom colid.
 
 // Require manager clearance.
@@ -1966,7 +1827,7 @@ $_BAR->redirectTo(lovd_getInstallURL() . 'columns/' . $zData['category'], 3);
 
 if (!empty($_PATH_ELEMENTS[2]) && ACTION == 'remove') {
     // URL: /columns/Variant/DNA?remove
-    // Drop specific custom column.
+    // Disable specific custom column.
     
     $aCol = $_PATH_ELEMENTS;
     unset($aCol[0]); // 'columns';
@@ -2240,13 +2101,15 @@ if (!empty($_PATH_ELEMENTS[2]) && ACTION == 'delete') {
     $sColumnID = implode('/', $aCol);
     $sCategory = $aCol[1];
 
-    $zData = $_DB->query('SELECT c.id, c.hgvs, c.head_column, ac.colid FROM ' . TABLE_COLS . ' AS c LEFT OUTER JOIN ' . TABLE_ACTIVE_COLS . ' AS ac ON (c.id = ac.colid) WHERE c.id = ?', array($sColumnID))->fetchAssoc();
+    $zData = $_DB->query('SELECT c.id, c.hgvs, c.head_column, ac.colid, c.created_by FROM ' . TABLE_COLS . ' AS c LEFT OUTER JOIN ' . TABLE_ACTIVE_COLS . ' AS ac ON (c.id = ac.colid) WHERE c.id = ?', array($sColumnID))->fetchAssoc();
 
     $sMessage = '';
     if (!$zData) {
         $sMessage = 'No such column!';
     } elseif ($zData['colid']) {
         $sMessage = 'Column is still active, disable it first!';
+    } elseif (!(int) $zData['created_by']) {
+        $sMessage = 'Only custom columns created by an LOVD user may be deleted from the system. This column however, is created by LOVD itself.';
     } elseif ($zData['hgvs']) {
         lovd_writeLog('Error', 'HackAttempt', 'Tried to remove HGVS column ' . $zData['id'] . ' (' . $zData['head_column'] . ')');
         $sMessage = 'Hack Attempt!';
@@ -2303,7 +2166,7 @@ if (!empty($_PATH_ELEMENTS[2]) && ACTION == 'delete') {
 
             require ROOT_PATH . 'inc-top.php';
             lovd_printHeader(PAGE_TITLE);
-            lovd_showInfoTable('Successfully deleted the column entry!', 'success');
+            lovd_showInfoTable('Successfully deleted the column ' . $sColumnID . '!', 'success');
 
             require ROOT_PATH . 'inc-bot.php';
             exit;
@@ -2327,10 +2190,10 @@ if (!empty($_PATH_ELEMENTS[2]) && ACTION == 'delete') {
     $aForm = array_merge(
                  array(
                         array('POST', '', '', '', '50%', '14', '50%'),
-                        array('Deleting this column', '', 'print', '<B>' . $sColumnID . '</B>'),
+                        array('Permanently deleting column', '', 'print', '<B>' . $sColumnID . '</B>'),
                         'skip',
                         array('Enter your password for authorization', '', 'password', 'password', 20),
-                        array('', '', 'submit', 'Delete column entry'),
+                        array('', '', 'submit', 'Delete column permanently'),
                       ));
     lovd_viewForm($aForm);
